+++ src/reducer/reducerHash.js
@@ -1,6 +1,8 @@
-import identity   from 'lodash.identity';
-import isFunction from 'lodash.isfunction';
-import verify     from '../util/verify';
+import identity          from 'lodash.identity';
+import isFunction        from 'lodash.isfunction';
+import verify            from '../util/verify';
+import makePatchableHOF  from '../util/makePatchableHOF';
+
 
 /**
  * Create a higher-order reducer by combining a set of sub-reducer
@@ -17,8 +19,13 @@ import verify     from '../util/verify';
  * {@tutorial conceptHash}), and additional examples can be found in
  * {@tutorial conceptJoin} and {@tutorial fullExample}.
  *
- * **NOTE**: Because this function is so central to the rudimentary aspects of
- * reduction, it is common to provide a value-added {@tutorial logExt}.
+ * **NOTE**: Because reducerHash is so central to the rudimentary
+ * aspect of reduction, it is a common practice to extend it,
+ * promoting a 
+ * [centralized reducer-based logging capability]{@tutorial logExt}, 
+ * with an ability to correlate logging levels to state changes
+ * *(providing a means to filter logs at a high level with minimal
+ * output)*.
  *
  * @param {ActionReducerHash} actionHandlers - a hash of reducer functions,
  * indexed by the standard redux action.type.
@@ -28,7 +35,7 @@ import verify     from '../util/verify';
  * 
  * @returns {reducerFn} a newly created reducer function (described above).
  */
-export default function reducerHash(actionHandlers, initialState) {
+function reducerHash(actionHandlers, initialState) {
 
   // validate params
   const check = verify.prefix('AstxReduxUtil.reducerHash() parameter violation: ');
@@ -65,8 +72,10 @@ export default function reducerHash(actionHandlers, initialState) {
     // execute the handler indexed by the action.type (or the identity pass-through)
     return locateHandler(action)(state, action, originalReducerState);
   };
-
 }
+
+// NOTE: JSDoc cannot handle an in-line "export default makePatchableHOF()" above, so we do it here
+export default makePatchableHOF(reducerHash);